requirements.txt
kivy
opencv-python
numpy
face_recognition
mysql-connector-python
Pillow